Started adding support for different EV3 sensors.
I'm not sure about my implementation of the RGB color sensor, but other than that, I just added placeholder variables for the other sensors. I'm not sure whether or not to put in the allocations for original NXT sensors or 3rd party sensors. I will work on getting the additional sensors working.